Bridge 3.1. 4. interfejs 5. refakor typu 6. przeniesienie 7. zmiana nazwy pakietu
