+++ src/main/java/edu/kis/vh/nursery/stack/IntArrayStack.java
@@ -1,33 +1,38 @@
 package edu.kis.vh.nursery.stack;
 
-public class IntArrayStack {
+public class IntArrayStack implements Stacker {
 
     private static final int MAX_STACK = 12;
     private static final int INIT_EMPTY_STACK = -1;
     private int[] NUMBERS = new int[MAX_STACK];
     private int total = INIT_EMPTY_STACK;
 
-    public void countIn(int in) {
+    @Override
+    public void push(int in) {
         if (!isFull())
             NUMBERS[++total] = in;
     }
 
-    public boolean callCheck() {
+    @Override
+    public boolean isEmpty() {
         return total == INIT_EMPTY_STACK;
     }
 
+    @Override
     public boolean isFull() {
         return total == (MAX_STACK -1);
     }
 
-    public int peekaboo() {
-        if (callCheck())
+    @Override
+    public int top() {
+        if (isEmpty())
             return -1;
         return NUMBERS[total];
     }
 
-    public int countOut() {
-        if (callCheck())
+    @Override
+    public int pop() {
+        if (isEmpty())
             return -1;
         return NUMBERS[total--];
     }
