+++ src/main/java/edu/kis/vh/nursery/stack/Stacker.java
@@ -1,0 +1,13 @@
+package edu.kis.vh.nursery.stack;
+
+public interface Stacker {
+    void push(int i);
+
+    boolean isEmpty();
+
+    boolean isFull();
+
+    int top();
+
+    int pop();
+}
